Refactor global.h: Change return type of awaitPressAnyKey()
Refactor the global.h file to change the return type of the awaitPressAnyKey() function from void to char. This change allows the function to return the key pressed by the user, providing more flexibility in handling user input.

Refactor login.h and screen.h: Add include for account.h

In the login.h and screen.h files, add the necessary include statement for the account.h header file. This inclusion is required to use the AccountList data structure in the loginScreen() and managerScreen() functions.

Refactor screen.h: Change function signature in controlSubScreen()

Modify the function signature of the controlSubScreen() function in the screen.h file to include the AccountList pointer as a parameter. This change allows the function to access and manipulate the account list when called.

Update main.c: Initialize account list in main()

In the main.c file, initialize the account list by calling the initializeAccountList() function. This initialization is necessary to create an empty list that can be used to store and manage account data throughout the program.

Refactor screen.c: Change function signature in controlSubScreen()

Modify the function signature of the controlSubScreen() function in the screen.c file to include the AccountList pointer as a parameter. This change allows the function to access and manipulate the account list when called.

Refactor login.c: Change function signature in loginScreen()

Modify the function signature of the loginScreen() function in the login.c file to include the AccountList pointer as a parameter. This change allows the function to access and manipulate the account list when called.

Update account.h: Add function definitions and validationCode()

In the account.h file, add function definitions for createAccount(), printAccount(), printAccountInPosition(), and validationCode(). Additionally, include the necessary header files and define the validationCode() function for validating account codes.

Refactor global.c: Change return type of awaitPressAnyKey()

Refactor the global.c file to change the return type of the awaitPressAnyKey() function from void to char. This change allows the function to return the key pressed by the user, providing more flexibility in handling user input.

Refactor account.c: Add printAccount() and printAccountInPosition()

In the account.c file, add the printAccount() and printAccountInPosition() functions. These functions are responsible for displaying the account details on the screen.